Changed tests structure, started escaping in docstrings
